Install plugins one-by-one in Dockerfile
Installing each plugin individually adds inconvenience when creating a
new plugin, but adds convenience when the plugins' source files are
changed. We have to remember to update the Dockerfile when a new plugin
is made, but the Dockerfile is only rebuilt when the plugin's 'setup.py'
file is changed. This is a similar pattern to how atlascope itself is
installed.