Use Nginx to proxy requests to MinIO from within the Docker container
When run in our Docker Compose configuration, Django knows that,
internally, it should reach out to minio:9000 to talk to MinIO. We tell
Django to link to localhost:9000 for files hosted by MinIO since
MinIO is attached to port 9000 of the host.

In normal circumstances, Django would not link to localhost:9000 for
these files, but a FQDN where the object store lives or the proxy to
that object store. In that case, the link is valid whether it is
requested from within the Docker container or not.

However, in our development environment case, localhost:9000 *is not*
valid from within the container, but *is* valid from within the host.
This means that Django can't access the files that it itself is linking
to.

This Nginx instance is launched as part of the entrypoint to the Docker
container. It listens locally on port 9000 and transparently proxies
those requests to MinIO.