Add menu for smoother launching of vis tools
Now, in order to launch a visualization service (currently only Glance is supported),
a user must make 2 clicks. First on the pin representing the 3D volume in context,
and then on the "Launch" link in the menu that pops up. This makes leaving Atlascope
to open another tool less jarring.